break into Header and Landing components and add tests
