+++ src/components/Header.test.js
@@ -1,0 +1,11 @@
+const { render, cleanup } = require('react-testing-library');
+import React from 'react';
+import Header from './Header';
+
+afterAll(cleanup);
+
+describe('Header', () => {
+  it('renders without crashing', () => {
+    render(<Header />);
+  });
+});
