+++ src/components/Landing.test.js
@@ -1,0 +1,11 @@
+const { render, cleanup } = require('react-testing-library');
+import React from 'react';
+import Landing from './Landing';
+
+afterAll(cleanup);
+
+describe('Header', () => {
+  it('renders without crashing', () => {
+    render(<Landing />);
+  });
+});
